// src/app/common/enums/crm-ref-type.enum.ts
export enum CrmRefType {
  GENDER = 1,
    NATIONALITY = '2',
    MILITARY_DEGREE = '4',
    CITIZENSHIP = '5',
    EDUCATION = '6',
    AWARD = '7',
    SECTOR = '8',
    COMPANY_LEVEL_TYPE = '9',
    PUNISH = '10',
    BRANCH_STATUS = '11',
    BRANCH_TYPE = '13',
    COMPANY_TYPE = '13',
    EMPLOYEE_LANGUAGE = '14',
    EMPLOYEE_DEGREE = '15',
    GIVEN_PLACE = '16',
    POSITION = '104',
    EMP_DOC_TYPES = '106',
    SCIENCE_LEVEL = '103',
    ACADEMIC = '102',
    RELATIVITY_TYPE = '105',
    EMP_ADDRESS_TYPE = '108',
    PROFESSION = '109',
    DISMISS_REASON_TYPE = '110',
    PERSON_DOC_TYPE = '203',
    CAR_FUEL_TYPE = '204',
    CAR_COLOR = '205',
    CAR_BRAND = '206',
    OWNERSHIP_DOC_TYPE = '207',
    CAR_BODY_TYPE = '208',
    CAR_GAS_PLACE = '209',
    CAR_CHECK_RESULT = '211',
    CAR_DRB_STATE = '224',
    CAR_CHECK_TYPE = '210',
    CAR_TEXNO_TYPES = '212',
    CAR_MODEL_TYPE  = '214',
    CAR_DOC_TYPE  = '218',
    DRIVER_TECHNICAL_INSPECTION_STATE = '219',
    CAR_TRUSTEE_STATE = '221',
    CAR_TRUSTEE_TYPE = '222',
    CAR_ADDRESS_TYPE = '201',
    DRIVER_CATEGORY = '301',
    DRIVER_CATEGORY_REASON = '308',
    DRIVER_CERTIFICATE_STATE = '302',
    Driver_ADDRESS_TYPE = '303',
    DRIVER_DEPRIVE_TYPE = '304',
    DRIVER_STATE = '305',
    DRIVER_CHECK_TYPE = '306',
    DRIVER_CHECK_RESULT = '307',
    DRIVER_SCHOLL = '309',
    CHECK_STATE = '702',
    CHECK_AMT_STATE = '704',
    CHECK_PERSON_STATE = '705',
}